add creation date, public modificator and DTO folder
