+++ backend/YourNews/YourNews.Shared/DTO/MainEntityDTO.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace YourNews.Shared.DTO
+{
+	public abstract class MainEntityDTO
+	{
+		public int Id { get; set; }
+		public DateTime CreatingDate { get; set; }
+	}
+}
