+++ backend/YourNews/YourNews.Shared/DTO/NewsToUserDTO.cs
@@ -1,0 +1,10 @@
+﻿
+
+namespace YourNews.Shared.DTO
+{
+	public class NewsToUserDTO: MainEntityDTO
+	{
+		public int NewsId { get; set; }
+		public int UserId { get; set; }
+	}
+}
