+++ backend/YourNews/YourNews.Shared/DTO/PaperDTO.cs
@@ -1,0 +1,9 @@
+﻿
+namespace YourNews.Shared.DTO
+{
+	public class PaperDTO: MainEntityDTO
+	{
+		public string Title { get; set; }
+		public string PublisherId { get; set; }
+	}
+}
