+++ backend/YourNews/YourNews.Shared/DTO/PublisherDTO.cs
@@ -1,0 +1,10 @@
+﻿
+
+namespace YourNews.Shared.DTO
+{
+	public class PublisherDTO
+	{
+		public string Name { get; set; }
+		public string Mail { get; set; }
+	}
+}
