--- backend/YourNews/YourNews.Shared/MainEntityDTO.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace YourNews.Shared
-{
-	public abstract class MainEntityDTO
-	{
-		public int Id { get; set; }
-	}
-}
