--- backend/YourNews/YourNews.Shared/NewsToUserDTO.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace YourNews.Shared
-{
-	class NewsToUserDTO: MainEntityDTO
-	{
-		public int NewsId { get; set; }
-		public int UserId { get; set; }
-	}
-}
