--- backend/YourNews/YourNews.Shared/PaperDTO.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace YourNews.Shared
-{
-	class PaperDTO: MainEntityDTO
-	{
-		public string Title { get; set; }
-		public string PublisherId { get; set; }
-	}
-}
