--- backend/YourNews/YourNews.Shared/PublisherDTO.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace YourNews.Shared
-{
-	class PublisherDTO
-	{
-		public string Name { get; set; }
-		public string Mail { get; set; }
-	}
-}
